data_1t/requirements.txt
psycopg2-binary==2.9.10
faker==33.0.0